Task 54 - Refactored in packages Dtos
