+++ src/main/java/it/aleph/omega/dto/tag/UpdateTagDto.java
@@ -1,4 +1,4 @@
-package it.aleph.omega.dto;
+package it.aleph.omega.dto.tag;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
