+++ src/main/java/it/aleph/omega/mapper/BookDtoMapper.java
@@ -1,13 +1,12 @@
 package it.aleph.omega.mapper;
 
-import it.aleph.omega.dto.BookDto;
-import it.aleph.omega.dto.CreateBookDto;
-import it.aleph.omega.dto.TagDto;
-import it.aleph.omega.dto.UpdateBookDto;
+import it.aleph.omega.dto.book.BookDto;
+import it.aleph.omega.dto.book.CreateBookDto;
+import it.aleph.omega.dto.tag.TagDto;
+import it.aleph.omega.dto.book.UpdateBookDto;
 import it.aleph.omega.model.Book;
 import it.aleph.omega.model.Tag;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 
 import java.util.List;
