+++ src/main/java/it/aleph/omega/mapper/TagDtoMapper.java
@@ -1,11 +1,10 @@
 package it.aleph.omega.mapper;
 
-import it.aleph.omega.dto.CreateTagDto;
-import it.aleph.omega.dto.TagDto;
-import it.aleph.omega.dto.UpdateTagDto;
+import it.aleph.omega.dto.tag.CreateTagDto;
+import it.aleph.omega.dto.tag.TagDto;
+import it.aleph.omega.dto.tag.UpdateTagDto;
 import it.aleph.omega.model.Tag;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
 
 import java.util.List;
